# pkg/service/docs/FairserviceAPI_swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "Fair Service API for managing fair data",
        "title": "Fair Service",
        "termsOfService": "http://swagger.io/terms/",
        "contact": {
            "name": "Jürgen Enge",
            "url": "https://ub.unibas.ch",
            "email": "juergen.enge@unibas.ch"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0"
    },
    "paths": {
...
        "/{partition}/abortupdate": {
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "ends aborts transaction for a source without removal of missing items",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "fairservice"
                ],
                "summary": "aborts update transaction",
                "operationId": "post-abort-update",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Partition",
                        "name": "partition",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "source to abort update",
                        "name": "source",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/fair.SourceData"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/service.FairResultMessage"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/service.FairResultMessage"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/service.FairResultMessage"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/service.FairResultMessage"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/service.FairResultMessage"
                        }
                    }
                }
            }
        },
        "/{partition}/endupdate": {
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "ends update transaction for a source with commit",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "fairservice"
                ],
                "summary": "ends update transaction",
                "operationId": "post-end-update",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Partition",
                        "name": "partition",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "source to end update",
                        "name": "source",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/fair.SourceData"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/service.FairResultMessage"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/service.FairResultMessage"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/service.FairResultMessage"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/service.FairResultMessage"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/service.FairResultMessage"
                        }
                    }
                }
            }
        },
        "/{partition}/item": {
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "creates a new item within a transaction",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "fairservice"
                ],
                "summary": "creates a new item",
                "operationId": "post-create-item",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Partition",
                        "name": "partition",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "source to abort update",
                        "name": "source",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/fair.ItemData"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/service.FairResultMessage"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/service.FairResultMessage"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/service.FairResultMessage"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/service.FairResultMessage"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/service.FairResultMessage"
                        }
                    }
                }
            }
        },
        "/{partition}/source": {
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "updates or creates source for a partition",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "fairservice"
                ],
                "summary": "update or create source",
                "operationId": "post-set-source",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Partition",
                        "name": "partition",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "source to set",
                        "name": "source",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/fair.Source"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/service.FairResultMessage"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/service.FairResultMessage"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/service.FairResultMessage"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/service.FairResultMessage"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/service.FairResultMessage"
                        }
                    }
                }
            }
        },
        "/{partition}/startupdate": {
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "starts update transaction for a source",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "fairservice"
                ],
                "summary": "starts update transaction",
                "operationId": "post-start-update",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Partition",
                        "name": "partition",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "source to start update",
                        "name": "source",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/fair.SourceData"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/service.FairResultMessage"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/service.FairResultMessage"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/service.FairResultMessage"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/service.FairResultMessage"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/service.FairResultMessage"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "fair.DataAccess": {
            "type": "string",
            "enum": [
                "public",
                "closed",
                "closed_data",
                "open_access"
            ],
            "x-enum-varnames": [
                "DataAccessPublic",
                "DataAccessClosed",
                "DataAccessClosedData",
                "DataAccessOpenAccess"
            ]
        },
...
        "fair.ItemData": {
            "type": "object",
            "properties": {
                "access": {
                    "$ref": "#/definitions/fair.DataAccess"
                },
                "catalog": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "datestamp": {
                    "type": "string"
                },
                "identifier": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "metadata": {
                    "$ref": "#/definitions/myfair.Core"
                },
                "partition": {
                    "type": "string"
                },
                "set": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "signature": {
                    "type": "string"
                },
                "source": {
                    "type": "string"
                },
                "status": {
                    "$ref": "#/definitions/fair.DataStatus"
                },
                "url": {
                    "type": "string"
                },
                "uuid": {
                    "type": "string"
                }
            }
        },
        "fair.Source": {
            "type": "object",
            "properties": {
                "description": {
                    "type": "string"
                },
                "detailURL": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "name": {
                    "type": "string"
                },
                "oaidomain": {
                    "type": "string"
                },
                "partition": {
                    "type": "string"
                }
            }
        },
        "fair.SourceData": {
            "type": "object",
            "properties": {
                "source": {
                    "type": "string"
                }
            }
        },
...
        "service.FairResultMessage": {
            "type": "object",
            "properties": {
                "message": {
                    "type": "string"
                },
                "status": {
                    "type": "string"
                }
            }
        }
    },
    "securityDefinitions": {
        "BearerAuth": {
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}